requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==1.3.5
pyyaml==6.0.1
python-dotenv==1.0.0
fake-useragent==1.1.3
lxml==4.9.3
tqdm==4.66.1